move close buttons to its own module
