+++ src/storage.js
@@ -91,3 +91,4 @@ export {
 	updateLocalStorage,
 	listLength
 };
+
